[FEATURE] Add support for uid
Previously, it was only possible to filter for table fields, if
they were defined in the TCA. However, it may be useful to be
able to filter by uid or other fields (which do not need a TCA
configuration).

This patch makes it possible to filter by

- uid (if the field uid exists in the table)
- and other fields (which exists in the table) for admins only

It was considered to add more support for other fields. This might
be made possible in the future by making the ALLOWED_FIELDS
configurable.

Resolves: #11